refactor: simplify dispatch and consolidate tests
- Drop EstablishedKinds() from FlowOnSuccessHandler; ManifestForOnSuccess
  is now a single switch table that both dispatch and the validator read.
- Inline purposeFlipTable into applyOutcomeFlip (two cases, no map needed).
- Inline handlerFor into runOnSuccess (one caller, one wired case).
- Replace establishedKinds (BFS-over-history union of manifests) with
  anyVisitedStepOnSuccess: today's only on_success establishes password,
  and the validator enforces upstream-collection separately.
- Merge flow_outcome_flip_test.go + flow_dispatch_test.go into
  flow_state_machine_test.go, and flow_definition_validator_adr017_test.go
  into flow_definition_validator_test.go.
