+++ internal/domain/flow_on_success_create_user.go
@@ -31,10 +31,6 @@ func NewFlowCreateUserHandler(ids idgen.Generator, users flowUserWriter, passwor
 
 var _ FlowOnSuccessHandler = (*FlowCreateUserHandler)(nil)
 
-func (h *FlowCreateUserHandler) EstablishedKinds() []FlowFieldChallenge {
-	return ManifestForOnSuccess(FlowOnSuccessCreateUser)
-}
-
 func (h *FlowCreateUserHandler) Handle(ctx context.Context, client database.QueryExecutor, in FlowOnSuccessInput) (FlowOnSuccessResult, error) {
 	identifierName, _, ok := findIdentifierField(in.Resolved.Fields, in.Fields)
 	if !ok {
