--- internal/domain/flow_outcome_flip_test.go
@@ -1,152 +1,0 @@
-package domain_test
-
-import (
-	"encoding/json"
-	"testing"
-
-	"github.com/stretchr/testify/assert"
-	"github.com/stretchr/testify/require"
-	"github.com/zitadel/nextgen/internal/domain"
-)
-
-func TestFlowStateMachine_Start_InitializesCurrentPurpose(t *testing.T) {
-	tests := []struct {
-		name    string
-		purpose domain.FlowDefinitionPurpose
-	}{
-		{"login", domain.FlowDefinitionPurposeLogin},
-		{"register", domain.FlowDefinitionPurposeRegister},
-		{"recovery", domain.FlowDefinitionPurposeRecovery},
-	}
-	for _, tc := range tests {
-		t.Run(tc.name, func(t *testing.T) {
-			w := newFlowTestWorld(t)
-			def := signupDefinition()
-			// signupDefinition only declares Register; register the other purposes
-			// so Start can resolve an entry step.
-			def.Purposes[domain.FlowDefinitionPurposeLogin] = "credentials"
-			def.Purposes[domain.FlowDefinitionPurposeRecovery] = "credentials"
-
-			result, err := w.sm.Start(t.Context(), nil, domain.FlowStartInput{
-				Definition:    def,
-				Purpose:       tc.purpose,
-				Session:       domain.FlowSessionRef{ID: "sess-1", Version: 1},
-				UserSchemaURL: defaultSchemaURL,
-			})
-			require.NoError(t, err)
-			assert.Equal(t, tc.purpose, result.State.Purpose)
-			assert.Equal(t, tc.purpose, result.State.CurrentPurpose,
-				"Start must initialise CurrentPurpose from Purpose")
-		})
-	}
-}
-
-func TestFlowStateMachine_FlipTable_LoginUserNotFoundFlipsToRegister(t *testing.T) {
-	w := newFlowTestWorld(t)
-	def := loginDefinition()
-
-	start, err := w.sm.Start(t.Context(), nil, domain.FlowStartInput{
-		Definition:    def,
-		Purpose:       domain.FlowDefinitionPurposeLogin,
-		Session:       domain.FlowSessionRef{ID: "sess-1", Version: 1},
-		UserSchemaURL: defaultSchemaURL,
-	})
-	require.NoError(t, err)
-	assert.Equal(t, domain.FlowDefinitionPurposeLogin, start.State.CurrentPurpose)
-
-	// Submitting an unknown identifier emits user_not_found.
-	result, err := w.sm.Process(t.Context(), nil, def, start.State, domain.FlowSubmitInput{
-		Action: domain.FlowActionSubmit,
-		Fields: map[string]any{
-			"email":    "ghost@example.com",
-			"password": "irrelevant",
-		},
-	})
-	require.NoError(t, err)
-	assert.Equal(t, domain.FlowDefinitionPurposeLogin, result.State.Purpose,
-		"Purpose stays pinned for telemetry / ACR")
-	assert.Equal(t, domain.FlowDefinitionPurposeRegister, result.State.CurrentPurpose,
-		"login + user_not_found flips CurrentPurpose to register")
-}
-
-func TestFlowStateMachine_FlipTable_RecoveryPassthrough(t *testing.T) {
-	w := newFlowTestWorld(t)
-	def := loginDefinition()
-	// Re-purpose the login definition as a recovery flow.
-	delete(def.Purposes, domain.FlowDefinitionPurposeLogin)
-	def.Purposes[domain.FlowDefinitionPurposeRecovery] = "credentials"
-
-	start, err := w.sm.Start(t.Context(), nil, domain.FlowStartInput{
-		Definition:    def,
-		Purpose:       domain.FlowDefinitionPurposeRecovery,
-		Session:       domain.FlowSessionRef{ID: "sess-1", Version: 1},
-		UserSchemaURL: defaultSchemaURL,
-	})
-	require.NoError(t, err)
-
-	result, err := w.sm.Process(t.Context(), nil, def, start.State, domain.FlowSubmitInput{
-		Action: domain.FlowActionSubmit,
-		Fields: map[string]any{
-			"email":    "ghost@example.com",
-			"password": "irrelevant",
-		},
-	})
-	require.NoError(t, err)
-	assert.Equal(t, domain.FlowDefinitionPurposeRecovery, result.State.CurrentPurpose,
-		"recovery never flips on identifier outcomes")
-}
-
-func TestFlowState_JSONRoundTrip_PreservesCurrentPurpose(t *testing.T) {
-	state := domain.FlowState{
-		ID:        "flow-1",
-		ProjectID: "proj-1",
-		FlowProgress: domain.FlowProgress{
-			DefinitionID:   "def-1",
-			Purpose:        domain.FlowDefinitionPurposeLogin,
-			CurrentPurpose: domain.FlowDefinitionPurposeRegister,
-			CurrentStep:    "credentials",
-			CollectedData:  map[string]any{},
-		},
-	}
-	payload, err := json.Marshal(state)
-	require.NoError(t, err)
-
-	var decoded domain.FlowState
-	require.NoError(t, json.Unmarshal(payload, &decoded))
-	assert.Equal(t, domain.FlowDefinitionPurposeLogin, decoded.Purpose)
-	assert.Equal(t, domain.FlowDefinitionPurposeRegister, decoded.CurrentPurpose,
-		"CurrentPurpose must survive a cookie round-trip")
-}
-
-func TestFlowState_JSONRoundTrip_PivotPushPopPreservesCurrentPurpose(t *testing.T) {
-	parent := domain.FlowProgress{
-		DefinitionID:   "def-parent",
-		Purpose:        domain.FlowDefinitionPurposeLogin,
-		CurrentPurpose: domain.FlowDefinitionPurposeRegister,
-		CurrentStep:    "parent-step",
-		CollectedData:  map[string]any{},
-	}
-	state := domain.FlowState{
-		ID:        "flow-1",
-		ProjectID: "proj-1",
-		FlowProgress: domain.FlowProgress{
-			DefinitionID:   "def-child",
-			Purpose:        domain.FlowDefinitionPurposeLogin,
-			CurrentPurpose: domain.FlowDefinitionPurposeLogin,
-			CurrentStep:    "child-step",
-			CollectedData:  map[string]any{},
-		},
-		PivotStack: []domain.FlowProgress{parent},
-	}
-
-	payload, err := json.Marshal(state)
-	require.NoError(t, err)
-	var decoded domain.FlowState
-	require.NoError(t, json.Unmarshal(payload, &decoded))
-
-	require.Len(t, decoded.PivotStack, 1)
-	assert.Equal(t, domain.FlowDefinitionPurposeRegister, decoded.PivotStack[0].CurrentPurpose,
-		"pivoted-parent CurrentPurpose must survive serialisation")
-	assert.Equal(t, domain.FlowDefinitionPurposeLogin, decoded.CurrentPurpose,
-		"active CurrentPurpose must survive serialisation")
-}
